+++ tests/test_arguments.py
@@ -4,6 +4,7 @@ from unittest import mock
 import pytest
 
 import click
+from click.utils import UNSET
 
 
 def test_nargs_star(runner):
@@ -174,6 +175,8 @@ def test_stdout_default(runner):
     result = runner.invoke(inout, [])
     assert not result.exception
     assert result.output == "Foo bar baz\n"
+    assert result.stdout == "Foo bar baz\n"
+    assert not result.stderr
 
 
 @pytest.mark.parametrize(
@@ -251,13 +254,37 @@ def test_missing_arg(runner):
     assert "Missing argument 'ARG'." in result.output
 
 
-def test_missing_argument_string_cast():
+@pytest.mark.parametrize(
+    ("value", "expect_missing", "processed_value"),
+    [
+        (UNSET, True, None),
+        (None, False, None),
+        # Default type of the argument is str, so everything is processed as strings.
+        ("", False, ""),
+        ("  ", False, "  "),
+        ("foo", False, "foo"),
+        ("12", False, "12"),
+        (12, False, "12"),
+        (12.1, False, "12.1"),
+        (list(), False, "[]"),
+        (tuple(), False, "()"),
+        (set(), False, "set()"),
+        (frozenset(), False, "frozenset()"),
+        (dict(), False, "{}"),
+    ],
+)
+def test_required_argument(value, expect_missing, processed_value):
     ctx = click.Context(click.Command(""))
+    argument = click.Argument(["a"], required=True)
 
-    with pytest.raises(click.MissingParameter) as excinfo:
-        click.Argument(["a"], required=True).process_value(ctx, None)
+    if expect_missing:
+        with pytest.raises(click.MissingParameter) as excinfo:
+            argument.process_value(ctx, value)
+        assert str(excinfo.value) == "Missing parameter: a"
 
-    assert str(excinfo.value) == "Missing parameter: a"
+    else:
+        value = argument.process_value(ctx, value)
+        assert value == processed_value
 
 
 def test_implicit_non_required(runner):
@@ -351,22 +378,110 @@ def test_nargs_specified_plus_star_ordering(runner):
     assert result.output.splitlines() == ["('a', 'b', 'c')", "d", "('e', 'f')"]
 
 
-def test_defaults_for_nargs(runner):
+@pytest.mark.parametrize(
+    ("args", "default", "expected"),
+    [
+        # Any iterable with the same number of arguments as nargs is valid.
+        [[], (1, 2), "(1, 2)"],
+        [[], (1.1, 2.2), "(1, 2)"],
+        [[], ("1", "2"), "(1, 2)"],
+        [[], (None, None), "(None, None)"],
+        [[], [1, 2], "(1, 2)"],
+        [[], {1, 2}, "(1, 2)"],
+        [[], frozenset([1, 2]), "(1, 2)"],
+        [[], {1: "a", 2: "b"}, "(1, 2)"],
+        # Empty iterable is valid if default is None.
+        [[], None, "None"],
+        # Arguments overrides the default.
+        [["3", "4"], (1, 2), "(3, 4)"],
+    ],
+)
+def test_good_defaults_for_nargs(runner, args, default, expected):
     @click.command()
-    @click.argument("a", nargs=2, type=int, default=(1, 2))
+    @click.argument("a", nargs=2, type=int, default=default)
     def cmd(a):
-        x, y = a
-        click.echo(x + y)
+        click.echo(repr(a))
+
+    result = runner.invoke(cmd, args)
+    assert result.output.strip() == expected
+
+
+@pytest.mark.parametrize(
+    ("default", "message"),
+    [
+        # Non-iterables defaults.
+        ["Yo", "Error: Invalid value for '[A]...': Value must be an iterable."],
+        ["", "Error: Invalid value for '[A]...': Value must be an iterable."],
+        [True, "Error: Invalid value for '[A]...': Value must be an iterable."],
+        [False, "Error: Invalid value for '[A]...': Value must be an iterable."],
+        [12, "Error: Invalid value for '[A]...': Value must be an iterable."],
+        [7.9, "Error: Invalid value for '[A]...': Value must be an iterable."],
+        # Generator default.
+        [(), "Error: Invalid value for '[A]...': Takes 2 values but 0 were given."],
+        # Unset default.
+        [UNSET, "Error: Missing argument 'A...'."],
+        # Tuples defaults with wrong length.
+        [
+            tuple(),
+            "Error: Invalid value for '[A]...': Takes 2 values but 0 were given.",
+        ],
+        [(1,), "Error: Invalid value for '[A]...': Takes 2 values but 1 was given."],
+        [
+            (1, 2, 3),
+            "Error: Invalid value for '[A]...': Takes 2 values but 3 were given.",
+        ],
+        # Lists defaults with wrong length.
+        [list(), "Error: Invalid value for '[A]...': Takes 2 values but 0 were given."],
+        [[1], "Error: Invalid value for '[A]...': Takes 2 values but 1 was given."],
+        [
+            [1, 2, 3],
+            "Error: Invalid value for '[A]...': Takes 2 values but 3 were given.",
+        ],
+        # Sets defaults with wrong length.
+        [set(), "Error: Invalid value for '[A]...': Takes 2 values but 0 were given."],
+        [
+            set([1]),
+            "Error: Invalid value for '[A]...': Takes 2 values but 1 was given.",
+        ],
+        [
+            set([1, 2, 3]),
+            "Error: Invalid value for '[A]...': Takes 2 values but 3 were given.",
+        ],
+        # Frozensets defaults with wrong length.
+        [
+            frozenset(),
+            "Error: Invalid value for '[A]...': Takes 2 values but 0 were given.",
+        ],
+        [
+            frozenset([1]),
+            "Error: Invalid value for '[A]...': Takes 2 values but 1 was given.",
+        ],
+        [
+            frozenset([1, 2, 3]),
+            "Error: Invalid value for '[A]...': Takes 2 values but 3 were given.",
+        ],
+        # Dictionaries defaults with wrong length.
+        [dict(), "Error: Invalid value for '[A]...': Takes 2 values but 0 were given."],
+        [
+            {1: "a"},
+            "Error: Invalid value for '[A]...': Takes 2 values but 1 was given.",
+        ],
+        [
+            {1: "a", 2: "b", 3: "c"},
+            "Error: Invalid value for '[A]...': Takes 2 values but 3 were given.",
+        ],
+    ],
+)
+def test_bad_defaults_for_nargs(runner, default, message):
+    """Some defaults are not valid when nargs is set."""
+
+    @click.command()
+    @click.argument("a", nargs=2, type=int, default=default)
+    def cmd(a):
+        click.echo(repr(a))
 
     result = runner.invoke(cmd, [])
-    assert result.output.strip() == "3"
-
-    result = runner.invoke(cmd, ["3", "4"])
-    assert result.output.strip() == "7"
-
-    result = runner.invoke(cmd, ["3"])
-    assert result.exception is not None
-    assert "Argument 'a' takes 2 values." in result.output
+    assert message in result.stderr
 
 
 def test_multiple_param_decls_not_allowed(runner):
@@ -381,12 +496,6 @@ def test_multiple_param_decls_not_allowed(runner):
 def test_multiple_not_allowed():
     with pytest.raises(TypeError, match="multiple"):
         click.Argument(["a"], multiple=True)
-
-
-@pytest.mark.parametrize("value", [(), ("a",), ("a", "b", "c")])
-def test_nargs_bad_default(runner, value):
-    with pytest.raises(ValueError, match="nargs=2"):
-        click.Argument(["a"], nargs=2, default=value)
 
 
 def test_subcommand_help(runner):