+++ tests/test_basic.py
@@ -7,6 +7,7 @@ from itertools import chain
 import pytest
 
 import click
+from click.utils import UNSET
 
 
 def test_basic_functionality(runner):
@@ -205,18 +206,28 @@ def test_float_option(runner, args, expect):
         assert result.exception is None
 
 
-@pytest.mark.parametrize("default", [True, False])
 @pytest.mark.parametrize(
-    ("args", "expect"), [(["--on"], True), (["--off"], False), ([], None)]
+    ("args", "default", "expect"),
+    [
+        (["--on"], True, True),
+        (["--on"], False, True),
+        (["--on"], None, True),
+        (["--on"], UNSET, True),
+        (["--off"], True, False),
+        (["--off"], False, False),
+        (["--off"], None, False),
+        (["--off"], UNSET, False),
+        ([], True, True),
+        ([], False, False),
+        ([], None, None),
+        ([], UNSET, False),
+    ],
 )
-def test_boolean_switch(runner, default, args, expect):
+def test_boolean_switch(runner, args, default, expect):
     @click.command()
     @click.option("--on/--off", default=default)
     def cli(on):
         return on
-
-    if expect is None:
-        expect = default
 
     result = runner.invoke(cli, args, standalone_mode=False)
     assert result.return_value is expect
@@ -229,6 +240,10 @@ def test_boolean_switch(runner, default, args, expect):
         (True, [], True),
         (False, ["--f"], True),
         (False, [], False),
+        # Boolean flags have a 3-states logic.
+        # See: https://github.com/pallets/click/issues/3024#issue-3285556668
+        (None, ["--f"], True),
+        (None, [], None),
     ),
 )
 def test_boolean_flag(runner, default, args, expect):
@@ -475,15 +490,25 @@ def test_choice_argument_none(runner):
     )
     def cli(method: str | None):
         assert isinstance(method, str) or method is None
-        click.echo(method)
+        click.echo(repr(method))
 
     result = runner.invoke(cli, ["not-none"])
     assert not result.exception
-    assert result.output == "not-none\n"
+    assert result.output == "'not-none'\n"
 
-    # None is not yet supported.
     result = runner.invoke(cli, ["none"])
+    assert not result.exception
+    assert result.output == "None\n"
+
+    result = runner.invoke(cli, [])
     assert result.exception
+    assert (
+        "Error: Missing argument '{not-none|none}'. "
+        "Choose from:\n\tnot-none,\n\tnone\n" in result.stderr
+    )
+
+    result = runner.invoke(cli, ["--help"])
+    assert result.output.startswith("Usage: cli [OPTIONS] {not-none|none}\n")
 
 
 def test_datetime_option_default(runner):