+++ CHANGES.rst
@@ -6,13 +6,23 @@ Version 8.2.x
 Unreleased
 
 -   show correct auto complete value for nargs option in combination with flag option :issue:`2813`
+Version 8.2.3
+-------------
+
+Unreleased
+
+-   Make first-class citizen of user-provided values to ``default`` and
+    ``flag_value`` arguments. This allows users to pass in values and have them
+    unaltered, by differentiating the absence of value from the value of absence. As
+    such `None` is now a valid value. :issue:`3024` :pr:`3030`
+-   Show correct auto complete value for nargs option in combination with flag option :issue:`2813`
 
 Version 8.2.2
 -------------
 
 Released 2025-07-31
 
--   Fix reconciliation of `default`, `flag_value` and `type` parameters for
+-   Fix reconciliation of ``default``, ``flag_value`` and ``type`` parameters for
     flag options, as well as parsing and normalization of environment variables.
     :issue:`2952` :pr:`2956`
 -   Fix typing issue in ``BadParameter`` and ``MissingParameter`` exceptions for the